organize helpers and project structure
