+++ test/constants.py
@@ -5,7 +5,7 @@ NNODES_ov_RLL10deg_CSne4 = 683
 NNODES_outCSne8 = 386
 NNODES_outCSne30 = 5402
 NNODES_outRLL1deg = 64442
-DATAVARS_outCSne30 = 3
+DATAVARS_outCSne30 = 2
 TRI_AREA = 1.047
 # 4*Pi is 12.56
 MESH30_AREA = 12.566
