+++ test/test_grid.py
@@ -8,6 +8,10 @@ from pathlib import Path
 
 import uxarray as ux
 
+from uxarray.grid.connectivity import _build_edge_node_connectivity, _build_face_edges_connectivity
+
+from uxarray.grid.coordinates import _populate_cartesian_xyz_coord, _populate_lonlat_coord
+
 try:
     import constants
 except ImportError:
@@ -429,7 +433,7 @@ class TestPopulateCoordinates(TestCase):
         verts_degree = np.stack((lon_deg, lat_deg), axis=1)
 
         vgrid = ux.open_grid(verts_degree, islatlon=False)
-        vgrid._populate_cartesian_xyz_coord()
+        _populate_cartesian_xyz_coord(vgrid)
 
         for i in range(0, vgrid.nMesh2_node):
             nt.assert_almost_equal(vgrid._ds["Mesh2_node_cart_x"].values[i],
@@ -470,7 +474,7 @@ class TestPopulateCoordinates(TestCase):
         verts_cart = np.stack((cart_x, cart_y, cart_z), axis=1)
 
         vgrid = ux.open_grid(verts_cart, islatlon=False)
-        vgrid._populate_lonlat_coord()
+        _populate_lonlat_coord(vgrid)
         # The connectivity in `__from_vert__()` will be formed in a reverse order
         lon_deg, lat_deg = zip(*reversed(list(zip(lon_deg, lat_deg))))
         for i in range(0, vgrid.nMesh2_node):
@@ -681,7 +685,7 @@ class TestConnectivity(TestCase):
         edge_nodes_expected = np.unique(edge_nodes_expected, axis=0)
 
         # construct edge nodes
-        mpas_grid_ux._build_edge_node_connectivity(repopulate=True)
+        _build_edge_node_connectivity(mpas_grid_ux, repopulate=True)
         edge_nodes_output = mpas_grid_ux._ds['Mesh2_edge_nodes'].values
 
         self.assertTrue(np.array_equal(edge_nodes_expected, edge_nodes_output))
@@ -704,7 +708,7 @@ class TestConnectivity(TestCase):
 
             mesh2_face_nodes = tgrid._ds["Mesh2_face_nodes"]
 
-            tgrid._build_face_edges_connectivity()
+            _build_face_edges_connectivity(tgrid)
             mesh2_face_edges = tgrid._ds.Mesh2_face_edges
             mesh2_edge_nodes = tgrid._ds.Mesh2_edge_nodes
 
@@ -739,7 +743,7 @@ class TestConnectivity(TestCase):
 
         mesh2_face_nodes = tgrid._ds["Mesh2_face_nodes"]
 
-        tgrid._build_face_edges_connectivity()
+        _build_face_edges_connectivity(tgrid)
         mesh2_face_edges = tgrid._ds.Mesh2_face_edges
         mesh2_edge_nodes = tgrid._ds.Mesh2_edge_nodes
 
@@ -762,7 +766,7 @@ class TestConnectivity(TestCase):
             self.f5_deg, self.f6_deg
         ]
         uds = ux.open_grid(verts)
-        uds._build_face_edges_connectivity()
+        _build_face_edges_connectivity(uds)
         n_face = len(uds._ds["Mesh2_face_edges"].values)
         n_node = uds.nMesh2_node
         n_edge = len(uds._ds["Mesh2_edge_nodes"].values)
