+++ uxarray/__init__.py
@@ -1,6 +1,6 @@
-from uxarray.utils.helpers import *
+# from uxarray.utils.helpers import *
 from uxarray.utils.constants import (INT_DTYPE, INT_FILL_VALUE)
-from uxarray.core.grid import Grid
+from uxarray.grid.grid import Grid
 from uxarray.core.api import (open_grid, open_dataset, open_mfdataset)
 from uxarray.core.dataarray import UxDataArray
 from uxarray.core.dataset import UxDataset
