+++ uxarray/io/_scrip.py
@@ -1,8 +1,8 @@
 import xarray as xr
 import numpy as np
 
-from uxarray.utils.helpers import grid_center_lat_lon, _replace_fill_values
-
+from uxarray.grid.utils import _replace_fill_values
+from uxarray.grid.coordinates import grid_center_lat_lon
 from uxarray.utils.constants import INT_DTYPE, INT_FILL_VALUE
